# Avoid usage of lazy-static