--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,5 +6,4 @@ edition = "2018"
 
 [dependencies]
 rand = "0.6.5"
-lazy_static = "1.3.0"
 rayon = "1.0.3"