Pass the two arrays as parameters for a slight performance improvement and to avoid a dependency.

```
before
  Time (mean ± σ):      3.287 s ±  0.016 s    [User: 26.088 s, System: 0.017 s]
  Range (min … max):    3.275 s …  3.329 s    10 runs

after
  Time (mean ± σ):      3.254 s ±  0.011 s    [User: 25.841 s, System: 0.012 s]
  Range (min … max):    3.243 s …  3.276 s    10 runs
```