@cfsamson 2019-03-13
Yes, that actually has a small effect. I actually tried this but had a hard time measuring it. Great!